+++ src/main/java/pascal/taie/AbstractWorldBuilder.java
@@ -73,7 +73,7 @@ public abstract class AbstractWorldBuilder implements WorldBuilder {
                             options.getAppClassPath().stream(),
                             options.getClassPath().stream())
                     .collect(Collectors.joining(File.pathSeparator));
-        } else { // when prependJVM is not set, we manually specify JRE jars
+        } else { // otherwise, we manually specify JRE jars
             // check existence of JREs
             File jreDir = new File(JREs);
             if (!jreDir.exists()) {